refactor: some factoring of code and in particular docker compose setup change
