+++ package.json
@@ -3,20 +3,25 @@
     "version": "1.0.0",
     "description": "Testing the LAPIS API, backed by SILO, designed for easily adding new test-cases.",
     "scripts": {
+        "pretest": "npm run dockerUp",
         "test": "vitest --reporter=html --reporter=default",
+        "posttest": "npm run dockerDown",
         "format": "prettier --write \"./**/*.{ts,tsx,json,astro,md,mdx,mjs,cjs}\"",
         "check-format": "prettier --check \"**/*.{ts,tsx,json,astro,md,mdx,mjs,cjs}\"",
         "check-types": "tsc --noEmit",
-        "check": "npm run check-format && npm run check-types"
+        "check": "npm run check-format && npm run check-types",
+        "dockerUp": "tsx dockerUp.ts",
+        "dockerDown": "tsx dockerDown.ts"
     },
     "type": "module",
     "devDependencies": {
         "@mongodb-js/zstd": "^1.2.1",
         "@types/node": "^22.5.4",
+        "@vitest/ui": "^2.0.5",
         "prettier": "^3.3.3",
+        "tsx": "^4.19.0",
         "typescript": "^5.5.4",
-        "vitest": "^2.0.5",
-        "@vitest/ui": "^2.0.5"
+        "vitest": "^2.0.5"
     },
     "engines": {
         "node": ">=22.0.0"
