+++ test.ts
@@ -1,119 +1,31 @@
-import { execSync } from 'child_process';
-import { afterAll, beforeAll, describe, expect, it } from 'vitest';
+import { describe, expect, it } from 'vitest';
 import * as path from 'path';
 import * as fs from 'fs';
-import { randomBytes } from 'node:crypto';
 import { decompress } from '@mongodb-js/zstd';
-import { TestCase } from './testCase.spec.js';
+import { TestCase } from './testCase.ts';
+import { getTestSets, TestSet } from './testSet.ts';
 
-// A salt for docker compose project names to not conflict with previous or parallel test runs
-const docker_project_salt = randomBytes(Math.ceil(2)).toString('hex').slice(0, 4);
-
-type TestSuite = {
-    path: string;
-    lapisPort: number;
-    siloPort: number;
-    testCases: TestCase[];
-};
-
-const PORT_BASE = 8000;
-
-const testsets: TestSuite[] = await Promise.all(
-    getTestsetDirectories().map(async (dir, index): Promise<TestSuite> => {
-        const lapisPort = PORT_BASE + index * 2;
-        const siloPort = PORT_BASE + index * 2 + 1;
-
-        const queriesDir = path.join(dir, 'queries');
-
-        const testCases: TestCase[] = await Promise.all(
-            fs
-                .readdirSync(queriesDir)
-                .filter((x) => x.endsWith('.query.ts'))
-                .map(async (file) => await loadTestObject(path.join(queriesDir, file))),
-        );
-
-        return { path: dir, lapisPort, siloPort, testCases };
-    }),
-);
+const testsets: TestSet[] = await getTestSets();
 
 console.log(
     'These are all the testsets that were identified with their corresponding LAPIS and SILO port numbers:',
-    testsets,
+    JSON.stringify(testsets),
 );
 
-testsets.map((testSuite) => {
-    const testBaseDir = path.basename(testSuite.path);
-    const testName = (testBaseDir + '_' + docker_project_salt).toLowerCase();
+testsets.map((testSet) => {
+    const testBaseDir = path.basename(testSet.path);
+    const testName = testBaseDir.toLowerCase();
 
-    const dataDir = path.join(testSuite.path, 'data');
-    const queriesDir = path.join(testSuite.path, 'queries');
-    const outputDir = path.join(testSuite.path, 'output');
-    if (!fs.existsSync(outputDir)) {
-        fs.mkdirSync(outputDir);
-    }
-
-    const dockerComposeEnv = [
-        'LAPIS_TAG=latest',
-        'SILO_TAG=latest',
-        `LAPIS_PORT=${testSuite.lapisPort}`,
-        `SILO_PORT=${testSuite.siloPort}`,
-        `TESTSET_DATA_FOLDER=${dataDir}`,
-        `TESTSET_OUTPUT_FOLDER=${outputDir}`,
-    ];
-
-    function dockerComposeUp() {
-        console.log(`Starting Docker Compose for ${testSuite.path}...`);
-
-        const dockerComposeUpCommand = [
-            ...dockerComposeEnv,
-            'docker',
-            'compose',
-            `--project-name=${testName}`,
-            '--progress=plain',
-            'up',
-            '--quiet-pull',
-            '--no-recreate',
-            '--detach',
-            '--wait',
-        ].join(' ');
-
-        console.log(dockerComposeUpCommand);
-        const execOptions = process.env.VERBOSE ? { stdio: 'inherit' as 'inherit' } : { stdio: 'ignore' as 'ignore' };
-        execSync(dockerComposeUpCommand, execOptions);
-    }
-
-    function dockerComposeDown() {
-        console.log(`Stopping Docker Compose for ${testSuite.path}...`);
-
-        const dockerComposeDownCommand = [
-            'docker',
-            'compose',
-            `--project-name=${testName}`,
-            '--progress=plain',
-            'down',
-        ].join(' ');
-
-        console.log(dockerComposeDownCommand);
-        const execOptions = process.env.VERBOSE ? { stdio: 'inherit' as 'inherit' } : { stdio: 'ignore' as 'ignore' };
-        execSync(dockerComposeDownCommand, execOptions);
-    }
+    const queriesDir = path.join(testSet.path, 'queries');
 
     describe(`Testset: ${testName}`, () => {
-        beforeAll(() => {
-            dockerComposeUp();
-        });
-
-        afterAll(() => {
-            dockerComposeDown();
-        });
-
-        testSuite.testCases.forEach((test) => {
-            itShouldValidateTestCase(test, queriesDir, testSuite.lapisPort);
+        testSet.testCases.forEach((test) => {
+            itShouldValidateTestCase(test, queriesDir, testSet.lapisPort);
         });
     });
 });
 
-async function itShouldValidateTestCase(testCase: TestCase, queriesDir: string, lapisPort: number) {
+function itShouldValidateTestCase(testCase: TestCase, queriesDir: string, lapisPort: number) {
     it(`should validate query from ${testCase.name}`, async () => {
         const url = `http://localhost:${lapisPort}${testCase.endpoint}`;
 
@@ -141,24 +53,6 @@ async function itShouldValidateTestCase(testCase: TestCase, queriesDir: string, 
             expect(actualResponse).to.deep.equal(expectedResponse);
         }
     });
-}
-
-function getTestsetDirectories(): string[] {
-    const testsetsPath = path.resolve(__dirname, 'testsets');
-    return fs
-        .readdirSync(testsetsPath)
-        .filter((name) => fs.lstatSync(path.join(testsetsPath, name)).isDirectory())
-        .map((name) => path.join(testsetsPath, name));
-}
-
-async function loadTestObject(filename: string): Promise<TestCase> {
-    try {
-        const module = await import(filename);
-        return module.default; // Access the default export
-    } catch (error) {
-        console.error(`Failed to load module: ${filename}`, error);
-        throw error;
-    }
 }
 
 async function getExpectedResponseString(file: string, compressed: boolean | undefined): Promise<string> {
